# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Event Registration API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Event Registration application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "paths": {
    "/events": {
      "post": {
        "summary": "Create a new event",
        "description": "Creates a new event in the system.",
        "tags": [
          "Events"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvent"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get all events",
        "description": "Retrieves a list of all events.",
        "tags": [
          "Events"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved events"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "summary": "Get event by ID",
        "description": "Retrieves a single event using its ID.",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the event"
          }
        ],
        "responses": {
          "200": {
            "description": "Event retrieved successfully"
          },
          "404": {
            "description": "Event not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update an event",
        "description": "Updates an existing event using its ID.",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the event"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Event not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete an event",
        "description": "Deletes an event using its ID.",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the event"
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted successfully"
          },
          "404": {
            "description": "Event not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Returns the health status of the API.",
        "tags": [
          "Health"
        ],
        "responses": {
          "200": {
            "description": "API is running successfully"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateEvent": {
        "type": "object",
        "required": [
          "name",
          "date",
          "location",
          "capacity"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "description": "Name of the event",
            "example": "Tech Conference"
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "Date of the event",
            "example": "2026-04-15"
          },
          "location": {
            "type": "string",
            "description": "Location of the event",
            "example": "Winnipeg"
          },
          "capacity": {
            "type": "integer",
            "minimum": 1,
            "description": "Maximum number of attendees",
            "example": 100
          }
        }
      },
      "UpdateEvent": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "description": "Updated name of the event",
            "example": "Updated Tech Conference"
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "Updated date of the event",
            "example": "2026-04-20"
          },
          "location": {
            "type": "string",
            "description": "Updated location of the event",
            "example": "Toronto"
          },
          "capacity": {
            "type": "integer",
            "minimum": 1,
            "description": "Updated maximum number of attendees",
            "example": 150
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type or code",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Name is required"
          }
        }
      }
    }
  },
  "tags": []
}